+++ src/Tests/Dfe.Complete.Api.Tests.Integration/Customizations/CustomWebApplicationDbContextFactoryCustomization.cs
@@ -131,7 +131,6 @@ namespace Dfe.Complete.Api.Tests.Integration.Customizations
                 services.AddCompleteApiClient<IProjectsClient, ProjectsClient>(config, client);
                 services.AddCompleteApiClient<IProjectGroupClient, ProjectGroupClient>(config, client);
                 services.AddCompleteApiClient<ITasksDataClient, TasksDataClient>(config, client);
-                services.AddCompleteApiClient<ICsvExportClient, CsvExportClient>(config, client);
                 services.AddCompleteApiClient<IUsersClient, UsersClient>(config, client);
                 services.AddCompleteApiClient<IServiceSupportClient, ServiceSupportClient>(config, client);
                 services.AddCompleteApiClient<IContactsClient, ContactsClient>(config, client);
@@ -143,7 +142,6 @@ namespace Dfe.Complete.Api.Tests.Integration.Customizations
                 fixture.Inject(serviceProvider.GetRequiredService<IProjectsClient>());
                 fixture.Inject(serviceProvider.GetRequiredService<IProjectGroupClient>());
                 fixture.Inject(serviceProvider.GetRequiredService<ITasksDataClient>());
-                fixture.Inject(serviceProvider.GetRequiredService<ICsvExportClient>());
                 fixture.Inject(serviceProvider.GetRequiredService<IUsersClient>());
                 fixture.Inject(serviceProvider.GetRequiredService<IServiceSupportClient>());
                 fixture.Inject(serviceProvider.GetRequiredService<IContactsClient>());