+++ CHANGELOG.md
@@ -16,6 +16,13 @@ Security in case of vulnerabilities.
 
 ## Unreleased
 
+### Changed
+- Throw exception when key_contacts record is missing on contact update tasks
+- Require incoming trust UKPRN on project completion
+
+### Fixed
+- external contacts list, show new trust name if incoming trust ukprn is null
+
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-12-01.1229...main) for everything awaiting release