+++ CHANGELOG.md
@@ -16,8 +16,13 @@ Security in case of vulnerabilities.
 
 ## Unreleased
 
+### Changed
+- Throw exception when key_contacts record is missing on contact update tasks
+- Require incoming trust UKPRN on project completion
+
 ### Fixed
-- Fixed view notes page allowing text to flow off page
+- external contacts list, show new trust name if incoming trust ukprn is null
+- Fix View notes flowing off page
 
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-12-01.1229...main) for everything awaiting release