+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Security in case of vulnerabilities.
 
 ### Fixed
 - external contacts list, show new trust name if incoming trust ukprn is null
+- Fix View notes flowing off page
 
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-12-01.1229...main) for everything awaiting release
@@ -71,6 +72,14 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 ## [1.29.0](https://github.com/DFE-Digital/complete-conversions-transfers-changes/releases/tag/production-2025-11-14.1133) - 2025-11-13
 
 ### Added
+- Added GOV.UK Notify email notifications for project creation and assignment events
+- Added 'Confirm the incoming trust has completed all actions' task
+- Added `Request a new URN and record for the academy` task for transfer projects.
+- Added `Trust modification order task` task for conversion projects.
+- Added `Delete project` functionality.
+- Added `Share the information about the opening' task for conversion project.
+- Added `Subleases` task
+- Added `Tenancy at will` task for conversion project.
 - Added `Confirm the school has completed all actions` task for conversion project.
 
 ### Fixed